+++ src/peigs/src/c/pdspgvx.c
@@ -420,7 +420,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
       *info = linfo;
       
     fprintf( stderr, " %s me = %d argument %d is a pointer to NULL. \n",
-             msg, me, -linfo );
+             msg, (int)me, (int)(-linfo) );
     xstop_( &linfo );
     return;
   }
@@ -464,7 +464,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
   if ( *info != 0 ) {
       linfo = *info;
       fprintf( stderr, " %s me = %d argument %d has an illegal value. \n",
-               msg, me, -linfo);
+               msg, (int)me, (int)(-linfo));
       xstop_( info );
       return;
   }
@@ -510,7 +510,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
   if ( *info != 0 ) {
       linfo = *info;
       fprintf( stderr, " %s me = %d argument %d contains a pointer to NULL. \n",
-               msg, me, -linfo);
+               msg, (int)me, (int)(-linfo));
       xstop_( info );
       return;
   }
@@ -529,7 +529,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
     *info = -8;
 
   if ( *info != 0 ) {
-      fprintf( stderr, " %s me = %d mapA,B,Z differ \n", msg, me );
+      fprintf( stderr, " %s me = %d mapA,B,Z differ \n", msg, (int)me );
       xstop_( info );
       return;
   }
@@ -554,7 +554,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
   
   if ( linfo != 0 ) {
     *info = linfo;
-    fprintf( stderr, " %s me = %d Insufficient workspace provided. \n", msg, me );
+    fprintf( stderr, " %s me = %d Insufficient workspace provided. \n", msg, (int)me );
     xstop_( info );
     return;
   }
@@ -707,7 +707,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
       choleski( &msize, vecB, mapB, i_scrat, d_scrat, info );
       
       if( *info != 0 ) {
-        fprintf(stderr, " %s me = %d choleski returned info = %d \n", msg, me, *info );
+        fprintf(stderr, " %s me = %d choleski returned info = %d \n", msg, (int)me, (int)(*info) );
         *info = 1;
       }
       
@@ -737,7 +737,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
 #endif
 
         if( *info != 0 ) {
-          fprintf(stderr, " %s me = %d inverseL returned info = %d \n", msg, me, *info );
+          fprintf(stderr, " %s me = %d inverseL returned info = %d \n", msg, (int)me, (int)(*info) );
           *info = 2;
         }
       
@@ -825,7 +825,7 @@ void pdspgvx( ifact, ivector, irange, n, vecA, mapA, vecB, mapB,
 #endif
 
     if( *info != 0 ) {
-      fprintf(stderr, " %s me = %d pdspevx returned info = %d \n", msg, me, *info );
+      fprintf(stderr, " %s me = %d pdspevx returned info = %d \n", msg, (int)me, (int)(*info) );
       *info = 3;
 	exit(-1);
     }
